added data structures for file upload
