--- src/HTTPRequestUtils.cpp
@@ -1,238 +1,0 @@
-#include "webserv.hpp"
-
-bool isOrigForm(std::string &requestTarget, int &queryStart)
-{
-	for (int i = 0; i < (int)requestTarget.length(); i++)
-	{
-		if (requestTarget[i] == '?')
-		{
-			queryStart = i;
-			return (true);
-		}
-	}
-	return (false);
-}
-
-bool isInvalidChar(const unsigned char &c)
-{
-	if ((c <= 31) || c == 127)
-		return (true);
-	return (false);
-}
-
-void skipRequestLine(const char *request, unsigned int &i)
-{
-	while (request[i])
-	{
-		if (hasCRLF(request, i, 0))
-		{
-			i += 2;
-			return;
-		}
-		i++;
-	}
-}
-
-void skipHeader(const char *request, unsigned int &i)
-{
-	while (request[i])
-	{
-		if (hasCRLF(request, i, 1))
-		{
-			i += 4; // skip "\r\n\r\n"
-			return;
-		}
-		i++;
-	}
-}
-
-bool isValidContentType(std::string type)
-{
-	if (type == "text/plain" || type == "text/html" || \
-	type.substr(0, 30) == "multipart/form-data; boundary=")
-		return (true);
-	return (false);
-}
-
-bool isValidHost(std::string host)
-{
-	(void)host;
-	return (true);
-}
-
-/*
-The combination of \\r\\n
-is as a standard way to denote the end of a line in HTTP headers.
-*/
-bool hasCRLF(const char *request, unsigned int &i, int mode)
-{
-	if (mode == 0)
-	{
-		if (request[i] && request[i + 1] && request[i] == '\r' && request[i + 1] == '\n')
-			return (true);
-		return (false);
-	}
-	if (mode == 1)
-	{
-		if (request[i] && request[i + 1] && request[i + 2] && request[i + 3] && request[i] == '\r' &&
-			request[i + 1] == '\n' && request[i + 2] == '\r' && request[i + 3] == '\n')
-			return (true);
-		return (false);
-	}
-	return (false);
-}
-
-std::string extractValue(std::string &variables, int &i)
-{
-	int startPos;
-
-	startPos = i++;
-	while (variables[i] && variables[i] != '&')
-	{
-		if (variables[i] == '=')
-			return ("");
-		i++;
-	}
-	if (variables[startPos] == '=')
-		startPos++;
-	if (variables[i] == '&' && !variables[i + 1])
-		return ("");
-	return (variables.substr(startPos, i - startPos));
-}
-
-std::string extractKey(std::string &variables, int &i, int startPos)
-{
-	if (i == 0 || variables[i] == '?')
-		return ("");
-	if (variables[startPos] == '&' && startPos != 0)
-		startPos++;
-	if (variables.substr(startPos, i - startPos).find('&') != std::string::npos ||
-		variables.substr(startPos, i - startPos).find('?') != std::string::npos)
-		return ("");
-	return (variables.substr(startPos, i - startPos));
-}
-
-std::string extractRequestTarget(const char *request, unsigned int &i)
-{
-	std::string requestTarget;
-	std::string string_request(request);
-	unsigned int start = i;
-
-	while (request[i] && request[i] != ' ' && !isInvalidChar(request[i]))
-		i++;
-	if (i > MAX_URI)
-		return ("");
-	requestTarget = string_request.substr(start, i - start);
-	return (requestTarget);
-}
-
-std::string extractVariables(std::string &requestTarget, bool &isOriginForm)
-{
-	int queryStart = 0;
-
-	if (requestTarget == "/")
-		return ("/");
-	isOriginForm = isOrigForm(requestTarget, queryStart);
-	if (isOriginForm)
-		return (requestTarget.substr(queryStart + 1, strlen(requestTarget.c_str()) - queryStart));
-	return (requestTarget);
-}
-
-std::string extractProtocolVersion(const char *request, unsigned int &i)
-{
-	std::string protocolVersion;
-	std::string string_request(request);
-	unsigned int start = i;
-
-	while (request[i] && request[i] != '\r' && !isInvalidChar(request[i]))
-		i++;
-	protocolVersion = string_request.substr(start, i - start);
-	if (protocolVersion == "HTTP/1.1")
-		return (protocolVersion);
-	return ("");
-}
-
-std::string extractMethod(const char *request, unsigned int &i)
-{
-	std::string method;
-	std::string string_request(request);
-
-	while (request[i] && request[i] != ' ' && !isInvalidChar(request[i]))
-		i++;
-	method = string_request.substr(0, i);
-	if (method == "GET" || method == "POST" || method == "DELETE")
-		return (method);
-	return ("");
-}
-
-std::string extractHeaderKey(const char *request, unsigned int &i)
-{
-	std::string string_request(request);
-	unsigned int start = i;
-
-	while (request[i] && request[i] != ':')
-	{
-		if (isInvalidChar(request[i]) || request[i] == ' ')
-			return ("");
-		i++;
-	}
-	return (string_request.substr(start, i - start));
-}
-
-std::string extractHeaderValue(const char *request, unsigned int &i)
-{
-	std::string string_request(request);
-	unsigned int start = i;
-
-	while (request[i] && request[i] != '\r')
-	{
-		if (isInvalidChar(request[i]))
-			return ("");
-		i++;
-	}
-	return (string_request.substr(start, i - start));
-}
-
-unsigned int extractLineLength(const char *request, unsigned int &i)
-{
-	std::string string_request(request);
-	unsigned int start = i;
-	int size = 0;
-
-	while (request[i] && request[i] != '\r')
-		i++;
-	if (!hasCRLF(request, i, 0))
-		return (-1);
-	size = hexToInt(string_request.substr(start, i - start));
-	if (size <= 0)
-		return (size);
-	i += 2; // skip '\r' and '\n'
-	return (size);
-}
-
-std::string extractLine(const char *request, unsigned int &i, const unsigned int &size)
-{
-	std::string string_request(request);
-	std::string line = string_request.substr(i, size);
-	i += size; // skip line
-	if (!hasCRLF(request, i, 0))
-		return ("");
-	i += 2; // skip '\r' and '\n'
-	return (line);
-}
-
-std::string extractUploadBoundary(std::string line)
-{
-	unsigned int start = 0;
-
-	for (unsigned int i = 0; i < line.length(); ++i){
-		if (line[i] == '='){
-			start = ++i;
-			while (i < line.length())
-				i++;
-			return (line.substr(start, i - start));
-		}
-	}
-	return ("");
-}
-
